Dashboard: Calendar
General:
Created functional navigation buttons that allows the user to change the month and year of the calendar.
Expanded calendar grid to seven-by-eight to accomidate months that span six weeks.
Fixed bugs.

Expanded:
Added "eventListeners.js" module that includes methods to add listeners for the click of the calendar buttons.
Modified "DOMStrings.js" module to include the DOMStrings for the calendar buttons.
Modified "time.js" module to rename classes, to enable the "displayMonthAndYear" method to accept arguments for the month and year, to add data attributes to the calendar headline, and to fix a bug that displayed 12:00 AM/PM as 00:00 AM/PM.
Modified "index.js" script to run the event listeners.
Modified "dashboard.scss" to rename several classes, to expand the calendar grid to seven-by-eight, and to fix a bug that incorrectly linked the clock face vector.
Modified "front-dashboard.ejs" to add identification attributes to the calendar buttons and to fix a bug that incorrectly linked the CSS stylesheet.