Dashboard: Navigation, Header, and Greeting
General:
Navigation and header buttons alter in appearance upon the user's hover over the components.
The greeting module has been modified to be a flex box to better react to viewport resizes and contains randomized quotes.
Fixed bugs.

Notes:
There may be browser compability issues because of the inclusion of the "-webkit-mask-image" CSS attribute.

Expanded:
Created a "greeting.js" module that contains a method that randomizes the choice of elements in the quote and author array.
Modified the "DOMStrings" module to include the DOMStrings for the quote and author.
Modified "index.js" script to run the greeting display method.
Modified "dashboard.scss" stylesheet to link button icons in the stylesheet, rather in the HTML document. Modified "dashboard.scss" to include hover effects over button icons, to dispaly the greeting module as a flexbox, and to fix a bug that caused the alarms module to be displayed incorrectly.
Modified "global.scss" to alter the default font size to 10px and added an additonal breakpoint at 1600px to adjust the font size.
Modified "front-dashboard.ejs" to accommodate the changes made to the button icons and greetings module.